Added start of work on making linux-based collectors.
kannel.conf is a config file that works for kannel and the Sierra Wireless
AirCard modems. This works for collection of SMS only

The skype_socket_controller is a qt app that binds the X11 interface of
linux skype to a socket so skype can be controlled with anything that
can use sockets.